Changed paths: M       niismooth.c M       niiskullstrip.cpp M       affineRegistration.cpp M       niiornlm.c M       niinucorrect.c M       niiaverage.c M       niiskullstrip.h M       reg-lib/_reg_tools.h M       reg-lib/_reg_blockMatching.h M       reg-lib/_reg_affineTransformation.h M       niilib.c M       niilib.h M       niiamap.c M       ptr4Dlib.cpp M       niisanlm.c
(1) Extended function read_nifti_float in niilib.c by reading-flag.
(2) Affine registration in niiskullstrip.cpp is now working!
(3) Enabled flags to disable debug printouts.


git-svn-id: svn://141.35.69.214/amap@191 3806922a-3115-4c2f-bdad-02e90158041f

+++ ptr4Dlib.cpp
@@ -167,7 +167,7 @@ double_to_ptr_4D(nifti_image *image, double *vol4d, int index_nt)
 			double_to_ptr_4D2<int>(image, vol4d, index_nt);
 			break;
 		case NIFTI_TYPE_FLOAT32:
-			double_to_ptr_4D2<double>(image, vol4d, index_nt);
+			double_to_ptr_4D2<float>(image, vol4d, index_nt);
 			break;
 		case NIFTI_TYPE_FLOAT64:
 			double_to_ptr_4D2<double>(image, vol4d, index_nt);
